[feat] flags and frames

+++ WeRace.Telemetry.Generator/Definition/EnumValue.cs
@@ -1,13 +1,15 @@
 namespace WeRace.Telemetry.Generator.Definition;
 
-public sealed class EnumValue : IWithValidation {
+public sealed class EnumValue : IWithValidation
+{
   [RequiredMember] public string Name { get; set; } = "";
 
-  [RequiredMember] public uint Value { get; set; }
+  [RequiredMember] public long Value { get; set; }
 
   [RequiredMember] public string Description { get; set; } = "";
 
-  public void Validate(ValidationContext context) {
+  public void Validate(ValidationContext context)
+  {
     if (Name.Trim().Length == 0)
       throw new ArgumentException("Name cannot be empty", nameof(Name));
     if (Description.Trim().Length == 0)
